#34 | Moved the enumerations to the new file
